FIX: login function error
Adjustment to the login functionality to resolve the issue of newly registered accounts being unable to log in.
Upon successful login, we now set session['username'] = username. In the /main access path, we have added a check to determine if the user is already logged in.
We have refined some of the error messages for the login and registration processes.